refactor(game): rename texture_extra -> texture_bms_bga
