+++ src/game/CMakeLists.txt
@@ -74,8 +74,8 @@ add_library(
     graphics/sprite_slider.cpp
     graphics/sprite_text.cpp
     graphics/sprite_video.cpp
+    graphics/texture_bms_bga.cpp
     graphics/texture_dynamic.cpp
-    graphics/texture_extra.cpp
     graphics/texture_video.cpp
     graphics/video.cpp
     input/input_mgr.cpp
