+++ src/game/graphics/texture_bms_bga.cpp
@@ -1,4 +1,4 @@
-#include "texture_extra.h"
+#include "texture_bms_bga.h"
 
 #include <algorithm>
 #include <memory>
